+++ src/Poc.Mobile.App/Views/IMainView.cs
@@ -1,5 +1,0 @@
-﻿namespace Poc.Mobile.App.Views
-{
-    //TODO this is code smell, should get rid of it
-    public interface IMainView { }
-}